+++ network/api/serializers.py
@@ -1,12 +1,12 @@
 from rest_framework import serializers
 
-from network.base.models import Data
+from network.base.models import Data, Station
 
 
 class DataSerializer(serializers.ModelSerializer):
     class Meta:
         model = Data
-        fields = ('id', 'start', 'end', 'observation', 'ground_station', 'payload')
+        fields = ('id', 'start', 'end', 'observation', 'ground_station', 'payload', 'payload_demode')
         read_only_fields = ['id', 'start', 'end', 'observation', 'ground_station']
 
 
@@ -15,11 +15,12 @@ class JobSerializer(serializers.ModelSerializer):
     tle0 = serializers.SerializerMethodField()
     tle1 = serializers.SerializerMethodField()
     tle2 = serializers.SerializerMethodField()
+    mode = serializers.SerializerMethodField()
 
     class Meta:
         model = Data
         fields = ('id', 'start', 'end', 'ground_station', 'tle0', 'tle1', 'tle2',
-                  'frequency')
+                  'frequency', 'mode')
 
     def get_frequency(self, obj):
         return obj.observation.transmitter.downlink_low
@@ -32,3 +33,18 @@ class JobSerializer(serializers.ModelSerializer):
 
     def get_tle2(self, obj):
         return obj.observation.tle.tle2
+
+    def get_mode(self, obj):
+        try:
+            return obj.observation.transmitter.mode.name
+        except:
+            return ''
+
+
+class SettingsSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Station
+        fields = ('uuid', 'name', 'alt', 'lat', 'lng', 'rig',
+                  'active', 'antenna', 'id', 'apikey')
+
+    apikey = serializers.CharField(read_only=True)