+++ network/base/admin.py
@@ -1,8 +1,13 @@
 from django.contrib import admin
 
 from network.base.models import (Antenna, Satellite, Station, Transmitter,
-                                 Observation, Data, Mode, Tle)
+                                 Observation, Data, Mode, Tle, Rig)
 
+
+@admin.register(Rig)
+class RigAdmin(admin.ModelAdmin):
+    list_display = ('name', 'rictld_number')
+    list_filter = ('name', )
 
 @admin.register(Mode)
 class ModeAdmin(admin.ModelAdmin):
@@ -27,11 +32,8 @@ class StationAdmin(admin.ModelAdmin):
 
 @admin.register(Satellite)
 class SatelliteAdmin(admin.ModelAdmin):
-    list_display = ('name', 'norad_cat_id', 'updated_date')
+    list_display = ('name', 'norad_cat_id')
     readonly_fields = ('name', 'names', 'image')
-
-    def updated_date(self, obj):
-        return obj.updated.strftime('%d.%m.%Y, %H:%M')
 
 
 @admin.register(Tle)