+++ network/static/js/observation_new.js
@@ -1,4 +1,4 @@
-$(function () {
+$( document ).ready( function(){
     var minstart = $('#datetimepicker-start').data('date-minstart');
     var maxrange = $('#datetimepicker-end').data('date-maxrange');
     $('#datetimepicker-start').datetimepicker();
@@ -12,18 +12,23 @@ $(function () {
         $('#datetimepicker-end').data('DateTimePicker').maxDate(moment.utc(e.date).add(maxrange, 'm'));
     });
 
-    $('#satellite-selection').bind('keyup change', function() {
-        var norad = $(this).find(':selected').data('norad');
+    function select_proper_transmitters(norad) {
         $('#transmitter-selection').prop('disabled', false);
         $('#transmitter-selection option').hide();
         $('#transmitter-selection option[data-satellite="'+norad+'"]').show().prop('selected', true);
 
         $('.tle').hide();
         $('.tle[data-norad="'+norad+'"]').show();
-    });
-});
+    }
 
-$( document ).ready( function(){
+    var norad = $(this).find(':selected').data('norad');
+    select_proper_transmitters(norad);
+
+    $('#satellite-selection').bind('keyup change', function() {
+        var norad = $(this).find(':selected').data('norad');
+        select_proper_transmitters(norad);
+    });
+
     $('#calculate-observation').click( function(){
         $('.calculation-result').show();
         $('#timeline').empty();
@@ -38,8 +43,9 @@ $( document ).ready( function(){
             beforeSend: function() { $('#loading').show(); }
         }).done(function(data) {
             $('#loading').hide();
-            if (data['error']) {
-                var error_msg = data['error'];
+            if (data.error) {
+                var error_msg = data.error;
+                console.log(data.error);
                 $('#windows-data').html('<span class="text-danger">' + error_msg + '</span>');
             } else {
                 var dc = 0; //Data counter
@@ -55,7 +61,7 @@ $( document ).ready( function(){
                         $('#windows-data').append('<input type="hidden" name="'+dc+'-starting_time" value="'+n.start+'">');
                         $('#windows-data').append('<input type="hidden" name="'+dc+'-ending_time" value="'+n.end+'">');
                         $('#windows-data').append('<input type="hidden" name="'+dc+'-station" value="'+k.id+'">');
-                        times.push({starting_time: starting_time, ending_time: ending_time})
+                        times.push({starting_time: starting_time, ending_time: ending_time});
                         dc = dc + 1;
                     });
                     suggested_data.push({label : label, times : times});
@@ -65,38 +71,42 @@ $( document ).ready( function(){
                 if (dc > 0) {
                     timeline_init(start_time, end_time, suggested_data);
                 } else {
-                    var error_msg = 'No Ground Station available for this observation window';
-                    $('#windows-data').html('<span class="text-danger">' + error_msg + '</span>');
+                    var empty_msg = 'No Ground Station available for this observation window';
+                    $('#windows-data').html('<span class="text-danger">' + empty_msg + '</span>');
                 }
             }
         });
     });
+
+    function timeline_init( start, end, payload ){
+        var start_time_timeline = moment.utc(start).valueOf();
+        var end_time_timeline = moment.utc(end).valueOf();
+
+        $('#timeline').empty();
+        $('.coloredDiv').css('background-color', 'transparent');
+        $('#name').empty();
+
+        var chart = d3.timeline()
+                      .stack()
+                      .beginning(start_time_timeline)
+                      .ending(end_time_timeline)
+                      .hover(function (d, i, datum) {
+                          var div = $('#hoverRes');
+                          var colors = chart.colors();
+                          div.find('.coloredDiv').css('background-color', colors(i));
+                          div.find('#name').text(datum.label);
+                      })
+                      .margin({left:140, right:10, top:0, bottom:50})
+                      .tickFormat({format: d3.time.format.utc('%H:%M'), tickTime: d3.time.minutes, tickInterval: 30, tickSize: 6});
+
+        var svg_width = 1140;
+        if (screen.width < 1200) { svg_width = 940; }
+        if (screen.width < 992) { svg_width = 720; }
+        if (screen.width < 768) { svg_width = screen.width - 30; }
+        var svg = d3.select('#timeline').append('svg').attr('width', svg_width)
+                    .datum(payload).call(chart);
+
+        $('#hoverRes').show();
+        $('#schedule-observation').removeAttr('disabled');
+    }
 });
-
-function timeline_init( start, end, payload ){
-    var start_time_timeline = moment.utc(start).valueOf();
-    var end_time_timeline = moment.utc(end).valueOf();
-
-    $('#timeline').empty();
-    $('.coloredDiv').css('background-color', 'transparent');
-    $('#name').empty();
-
-    var chart = d3.timeline()
-                  .stack()
-                  .beginning(start_time_timeline)
-                  .ending(end_time_timeline)
-                  .hover(function (d, i, datum) {
-                      var div = $('#hoverRes');
-                      var colors = chart.colors();
-                      div.find('.coloredDiv').css('background-color', colors(i))
-                      div.find('#name').text(datum.label);
-                  })
-                  .margin({left:140, right:10, top:0, bottom:50})
-                  .tickFormat({format: d3.time.format('%H:%M'), tickTime: d3.time.minutes, tickInterval: 30, tickSize: 6});
-
-    var svg = d3.select('#timeline').append('svg').attr('width', 1140)
-                .datum(payload).call(chart);
-
-    $('#hoverRes').show();
-    $('#schedule-observation').removeAttr('disabled');
-}